+++ py/desimodel/io.py
@@ -7,7 +7,6 @@ desimodel.io
 I/O utility functions for files in desimodel.
 """
 import os
-import sys
 import re
 import warnings
 from datetime import datetime
@@ -29,12 +28,17 @@ def load_throughput(channel):
     ----------
     channel : {'b', 'r', 'z'}
         Spectrograph channel.
+
+    Returns
+    -------
+    Throughput
+        A specter throughput object.
     """
     import specter.throughput
     channel = channel.lower()
     global _thru
     if channel not in _thru:
-        thrufile = os.path.join(os.environ['DESIMODEL'],'data','throughput','thru-{0}.fits'.format(channel))
+        thrufile = findfile('throughput/thru-{0}.fits'.format(channel))
         _thru[channel] = specter.throughput.load_throughput(thrufile)
     return _thru[channel]
 #
@@ -48,12 +52,17 @@ def load_psf(channel):
     ----------
     channel : {'b', 'r', 'z'}
         Spectrograph channel.
+
+    Returns
+    -------
+    PSF
+        A specter PSF object.
     """
     import specter.psf
     channel = channel.lower()
     global _psf
     if channel not in _psf:
-        psffile = os.path.join(os.environ['DESIMODEL'],'data','specpsf','psf-{0}.fits'.format(channel))
+        psffile = findfile('specpsf/psf-{0}.fits'.format(channel))
         _psf[channel] = specter.psf.load_psf(psffile)
     return _psf[channel]
 #
@@ -61,11 +70,16 @@ def load_psf(channel):
 #
 _params = None
 def load_desiparams():
-    """Returns DESI parameter dictionary loaded from desimodel/data/desi.yaml.
+    """Returns DESI parameter dictionary loaded from ``$DESIMODEL/data/desi.yaml``.
+
+    Returns
+    -------
+    :class:`dict`
+        The parameters read from the YAML file.
     """
     global _params
     if _params is None:
-        desiparamsfile = os.path.join(os.environ['DESIMODEL'],'data','desi.yaml')
+        desiparamsfile = findfile('desi.yaml')
         with open(desiparamsfile) as par:
             _params = yaml.safe_load(par)
 
@@ -90,24 +104,33 @@ def load_desiparams():
 # Added and still needs to be committed and pushed to desihub
 _gfa = None
 def load_gfa():
-    """Returns GFA table from desimodel/data/focalplane/gfa.ecsv"""
+    """Returns GFA table from ``$DESIMODEL/data/focalplane/gfa.ecsv``.
+
+    Returns
+    -------
+    :class:`~astropy.table.Table`
+        The data from the ECSV file.
+    """
     global _gfa
-    from astropy.table import Table
-    # os is imported already in the desimodel io.py
-    import os
     if _gfa is None:
-        gfaFile = os.path.join(os.environ['DESIMODEL'], 'data', 'focalplane', 'gfa.ecsv')
-        _gfa = Table.read(gfaFile, format = 'ascii.ecsv')
+        gfaFile = findfile('focalplane/gfa.ecsv')
+        _gfa = Table.read(gfaFile, format='ascii.ecsv')
     return _gfa
 #
 #
 #
 _deviceloc = None
 def load_deviceloc():
+    """Returns a table from ``$DESIMODEL/data/focalplane/fiberpos-all.fits``.
+
+    Returns
+    -------
+    :class:`~astropy.table.Table`
+        The data from the FITS file, with columns converted to uppercase.
+    """
     global _deviceloc
-    from astropy.table import Table
     if _deviceloc is None:
-        fiberposfile = os.path.join(os.environ['DESIMODEL'],'data','focalplane','fiberpos-all.fits')
+        fiberposfile = findfile('focalplane/fiberpos-all.fits')
         _deviceloc = Table.read(fiberposfile)
 
     #- Convert to upper case if needed
@@ -120,12 +143,16 @@ def load_deviceloc():
 
 _fiberpos = None
 def load_fiberpos():
-    """Returns fiberpos table from desimodel/data/focalplane/fiberpos.fits.
+    """Returns fiberpos table from ``$DESIMODEL/data/focalplane/fiberpos.fits``.
+
+    Returns
+    -------
+    :class:`~astropy.table.Table`
+        The data from the FITS file, sorted by ``FIBER``.
     """
     global _fiberpos
-    from astropy.table import Table
     if _fiberpos is None:
-        fiberposfile = os.path.join(os.environ['DESIMODEL'],'data','focalplane','fiberpos.fits')
+        fiberposfile = findfile('focalplane/fiberpos.fits')
         _fiberpos = Table.read(fiberposfile)
         _fiberpos.sort('FIBER')
         #- Convert to upper case if needed
@@ -154,50 +181,74 @@ def load_fiberpos():
 #
 _tiles = dict()
 def load_tiles(onlydesi=True, extra=False, tilesfile=None, cache=True):
-    """Return DESI tiles structure from desimodel/data/footprint/desi-tiles.fits.
+    """Return DESI tiles structure from ``$DESIMODEL/data/footprint/desi-tiles.fits``.
 
     Parameters
     ----------
-    onlydesi : :class:`bool` (default True)
+    onlydesi : :class:`bool`, optional
         If ``True``, trim to just the tiles in the DESI footprint.
-    extra : :class:`bool`, (default False)
-        If ``True``, include extra layers with PROGRAM='EXTRA'.
-    tilesfile : (str)
-        Name of tiles file to load; or None for default.
-        Without path, look in $DESIMODEL/data/footprint, otherwise load file.
-    cache : :class:`bool`, (default True)
-        Use cache of tiles data.
+    extra : :class:`bool`, optional
+        If ``True``, include extra layers with ``PROGRAM='EXTRA'``.
+    tilesfile : :class:`str`, optional
+        Name of tiles file to load; or None for default.  See Notes for
+        details.
+    cache : :class:`bool`, optional
+        If ``False``, force reload of data from tiles file, instead of
+        using cached values.
+
+    Returns
+    -------
+    :class:`~astropy.io.fits.FITS_rec`
+        The data table portion of the FITS file.
+
+    Raises
+    ------
+    :exc:`FileNotFoundError`
+        If the value of `tilesfile` does not exist.
+
+    Notes
+    -----
+    Keyword-based environment variable expansion is performed on the `tilesfile`
+    value, so *e.g.*::
+
+        tiles = load_tiles(tilesfile='{HOME}/my-tiles.fits')
+
+    will be expanded with the value of :envvar:`HOME`.
+
+    If the parameter `tilesfile` is set, this function uses the following
+    search method:
+
+    1. If the value includes an explicit path, even ``./``, use that file.
+    2. If the value does *not* include an explicit path, *and* the file name
+       is identical to a file in ``$DESIMODEL/data/footprint/``, use the
+       file in ``$DESIMODEL/data/footprint/`` and issue a warning.
+    3. If no matching file can be found at all, raise an exception.
     """
     global _tiles
 
     if tilesfile is None:
         # Use the default
-        tilesfile = os.path.join(
-            os.environ['DESIMODEL'], 'data', 'footprint', 'desi-tiles.fits')
+        tilesfile = findfile('footprint/desi-tiles.fits')
     else:
         # If full path isn't included, check local vs $DESIMODEL/data/footprint
         tilepath, filename = os.path.split(tilesfile)
         if tilepath == '':
             have_local = os.path.isfile(tilesfile)
-            checkfile = os.path.join(os.environ['DESIMODEL'],
-                                     'data', 'footprint', tilesfile)
+            checkfile = findfile(os.path.join('footprint', tilesfile))
             have_dmdata = os.path.isfile(checkfile)
             if have_dmdata:
                 if have_local:
-                    msg = '$DESIMODEL/data/footprint/{} is shadowed by a local'\
-                          ' file. Choosing $DESIMODEL file.'\
-                          ' Use tilesfile="./{}" if you want the local copy'\
-                          ' instead'.format(tilesfile, tilesfile)
+                    msg = ('$DESIMODEL/data/footprint/{0} is shadowed by a local' +
+                           ' file. Choosing $DESIMODEL file.' +
+                           ' Use tilesfile="./{0}" if you want the local copy' +
+                           ' instead.').format(tilesfile)
                     warnings.warn(msg)
                 tilesfile = checkfile
 
             if not (have_local or have_dmdata):
-                msg = 'File "{}" does not exist locally or in '\
-                      '$DESIMODEL/data/footprint/'.format(tilesfile)
-                if sys.version_info.major == 2:
-                    raise IOError(msg)
-                else:
-                    raise FileNotFoundError(msg)
+                msg = ('File "{}" does not exist locally or in ' +
+                       '$DESIMODEL/data/footprint/!').format(tilesfile)
+                raise FileNotFoundError(msg)
 
     #- standarize path location
     tilesfile = os.path.abspath(tilesfile.format(**os.environ))
@@ -217,7 +268,7 @@ def load_tiles(onlydesi=True, extra=False, tilesfile=None, cache=True):
 
         #- Check for out-of-date tiles file
         if np.issubdtype(tiledata['OBSCONDITIONS'].dtype, np.unsignedinteger):
-            warnings.warn('old desi-tiles.fits with uint16 OBSCONDITIONS; please update your $DESIMODEL checkout', DeprecationWarning)
+            warnings.warn('Old desi-tiles.fits with uint16 OBSCONDITIONS; please update your $DESIMODEL checkout.', DeprecationWarning)
 
         #- load cache for next time
         if cache:
@@ -239,14 +290,32 @@ def load_tiles(onlydesi=True, extra=False, tilesfile=None, cache=True):
 
 _platescale = None
 def load_platescale():
-    '''
-    Loads platescale.txt, returning structured array with columns
+    """Loads platescale.txt.
 
-        radius: radius from center of focal plane [mm]
-        theta: radial angle that has a centroid at this radius [deg]
-        radial_platescale: Meridional (radial) plate scale [um/arcsec]
-        az_platescale: Sagittal (azimuthal) plate scale [um/arcsec]
-    '''
+    Returns
+    -------
+    :class:`~numpy.recarray`
+        The data table read from the file.
+
+    Notes
+    -----
+    The returned object has these columns:
+
+    radius
+        Radius from center of focal plane [mm].
+
+    theta
+        Radial angle that has a centroid at this radius [deg].
+
+    radial_platescale
+        Meridional (radial) plate scale [um/arcsec].
+
+    az_platescale:
+        Sagittal (azimuthal) plate scale [um/arcsec].
+
+    arclength:
+        Unknown description.
+    """
     global _platescale
     if _platescale is not None:
         return _platescale
@@ -267,16 +336,20 @@ _focalplane = None
 def load_focalplane(time=None):
     """Load the focalplane state that is valid for the given time.
 
-    Options:
-        time (datetime):  The time to query. default to current time.
+    Parameters
+    ----------
+    time : :class:`~datetime.datetime`
+        The time to query. Default to current time
+        (:meth:`~datetime.datetime.now`).
 
-    Returns:
-        (tuple):  The (FP layout, exclusion polygons, state, time string).
-            The FP layout is a Table.  The exclusion polygons are a dictionary
-            indexed by names that are referenced in the state.  The state
-            is a Table.  The time string is the resulting UTC ISO format
-            time string used by the lookup.
-
+    Returns
+    -------
+    :class:`tuple`
+        A tuple of (FP layout, exclusion polygons, state, time string).
+        The FP layout is a Table.  The exclusion polygons are a dictionary
+        indexed by names that are referenced in the state.  The state
+        is a Table.  The time string is the resulting UTC ISO format
+        time string used by the lookup.
     """
     if time is None:
         time = datetime.now()
@@ -406,9 +479,8 @@ def load_focalplane(time=None):
 
 
 def reset_cache():
-    '''Reset I/O cache'''
-    global _thru, _psf, _params, _gfa, _fiberpos, _tiles, _platescale,\
-        _focalplane
+    '''Reset I/O cache.'''
+    global _thru, _psf, _params, _gfa, _fiberpos, _tiles, _platescale, _focalplane
     _thru = dict()
     _psf = dict()
     _params = None
@@ -418,17 +490,22 @@ def reset_cache():
     _platescale = None
     _focalplane = None
 
+
 def load_target_info():
-    '''
-    Loads data/targets/targets.yaml and returns the nested dictionary
+    '''Loads data/targets/targets.yaml and returns the nested dictionary.
 
     This is primarily syntactic sugar to avoid end users constructing
-    paths and filenames by hand (which e.g. broke when targets.dat was
-    renamed to targets.yaml)
+    paths and filenames by hand (which *e.g.* broke when targets.dat was
+    renamed to targets.yaml).
+
+    Returns
+    -------
+    :class:`dict`
+        The dictionary read from the YAML file.
     '''
-    targetsfile = os.path.join(datadir(),'targets','targets.yaml')
+    targetsfile = findfile('targets/targets.yaml')
     if not os.path.exists(targetsfile):
-        targetsfile = os.path.join(datadir(),'targets','targets.dat')
+        targetsfile = findfile('targets/targets.dat')
 
     with open(targetsfile) as fx:
         data = yaml.safe_load(fx)
@@ -436,17 +513,19 @@ def load_target_info():
     return data
 
 def load_pixweight(nside, pixmap=None):
-    '''
-    Loads desimodel/data/footprint/desi-healpix-weights.fits
+    '''Loads ``$DESIMODEL/data/footprint/desi-healpix-weights.fits``.
 
-    Args:
-        nside: after loading, the array will be resampled to the
-            passed HEALPix nside
+    Parameters
+    ----------
+    nside : :class:`int`
+        After loading, the array will be resampled to the passed HEALPix `nside`.
+    pixmap : :class:`~astropy.io.fits.FITS_rec`, optional
+        Input pixel weight map, already read from a weights file.
 
-    Options:
-        pixmap: input pixel weight map (optional, defaults to None)
-
-    Returns healpix weight map for the DESI footprint at the requested nside
+    Returns
+    -------
+    Weight
+        HEALPix weight map for the DESI footprint at the requested `nside`.
     '''
     import healpy as hp
 
@@ -454,7 +533,7 @@ def load_pixweight(nside, pixmap=None):
         log.debug('Using input pixel weight map of length {}.'.format(len(pixmap)))
     else:
         #ADM read in the standard pixel weights file
-        pixfile = os.path.join(os.environ['DESIMODEL'],'data','footprint','desi-healpix-weights.fits')
+        pixfile = findfile('footprint/desi-healpix-weights.fits')
         with fits.open(pixfile) as hdulist:
             pixmap = hdulist[0].data
 
@@ -469,21 +548,31 @@ def load_pixweight(nside, pixmap=None):
     return hp.pixelfunc.ud_grade(pixmap,nside,order_in='NESTED',order_out='NESTED')
 
 def findfile(filename):
-    '''
-    Return full path to data file $DESIMODEL/data/filename
+    """Return full path to data file ``$DESIMODEL/data/filename``.
 
-    Note: this is a precursor for a potential future refactor where
-    desimodel data would be installed with the package and $DESIMODEL
+    Parameters
+    ----------
+    filename : :class:`str`
+        Name of the file, relative to the desimodel data directory.
+
+    Returns
+    -------
+    :class:`str`
+        The full path.
+
+    Notes
+    -----
+    This is a precursor for a potential future refactor where
+    desimodel data would be installed with the package and :envvar:`DESIMODEL`
     would become an optional override.
-    '''
+    """
     return os.path.join(datadir(), filename)
 
 def datadir():
-    '''
-    Returns location to desimodel data
+    """Returns location to desimodel data.
 
-    if set, $DESIMODEL overrides data installed with the package
-    '''
+    If set, :envvar:`DESIMODEL` overrides data installed with the package.
+    """
     if 'DESIMODEL' in os.environ:
         return os.path.abspath(os.path.join(os.environ['DESIMODEL'], 'data'))
     else: